test(cli): split format tests
